correction du style des form

+++ src/main/java/com/kobe/warehouse/service/facturation/dto/TiersPayantDossierFactureDto.java
@@ -1,3 +1,3 @@
 package com.kobe.warehouse.service.facturation.dto;
 
-public record TiersPayantDossierFactureDto(Long id, String name, Integer totalAmount, Long factureItemCount) {}
+public record TiersPayantDossierFactureDto(Integer id, String name, Integer totalAmount, Long factureItemCount) {}
